requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
spacy==3.6.1
nltk==3.8.1
sentence-transformers==2.2.2
transformers==4.33.2
torch==2.0.1
faiss-cpu==1.7.4
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
xgboost==1.7.6
lightgbm==4.0.0
matplotlib==3.7.2
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2
pytest==7.4.0
huggingface-hub==0.16.4
